@@ -1,24 +1,23 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk" ToolsVersion="15.0">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-    <LangVersion>latest</LangVersion>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="NLog" Version="4.5.11" />
+    <PackageReference Include="NLog" Version="4.6.3" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\refs\Omnix\src\Omnix.Base\Omnix.Base.csproj" />
-    <ProjectReference Include="..\..\refs\Omnix\src\Omnix.Collections\Omnix.Collections.csproj" />
-    <ProjectReference Include="..\..\refs\Omnix\src\Omnix.Configuration\Omnix.Configuration.csproj" />
-    <ProjectReference Include="..\..\refs\Omnix\src\Omnix.Correction\Omnix.Correction.csproj" />
-    <ProjectReference Include="..\..\refs\Omnix\src\Omnix.Io\Omnix.Io.csproj" />
-    <ProjectReference Include="..\..\refs\Omnix\src\Omnix.Network.Connection\Omnix.Network.Connection.csproj" />
-    <ProjectReference Include="..\..\refs\Omnix\src\Omnix.Network\Omnix.Network.csproj" />
-    <ProjectReference Include="..\..\refs\Omnix\src\Omnix.Serialization.RocketPack\Omnix.Serialization.RocketPack.csproj" />
-    <ProjectReference Include="..\..\refs\Omnix\src\Omnix.Serialization\Omnix.Serialization.csproj" />
+    <ProjectReference Include="..\..\refs\omnix\src\Omnix.Base\Omnix.Base.csproj" />
+    <ProjectReference Include="..\..\refs\omnix\src\Omnix.Collections\Omnix.Collections.csproj" />
+    <ProjectReference Include="..\..\refs\omnix\src\Omnix.Configuration\Omnix.Configuration.csproj" />
+    <ProjectReference Include="..\..\refs\omnix\src\Omnix.Correction\Omnix.Correction.csproj" />
+    <ProjectReference Include="..\..\refs\omnix\src\Omnix.Io\Omnix.Io.csproj" />
+    <ProjectReference Include="..\..\refs\omnix\src\Omnix.Network.Connection\Omnix.Network.Connection.csproj" />
+    <ProjectReference Include="..\..\refs\omnix\src\Omnix.Network\Omnix.Network.csproj" />
+    <ProjectReference Include="..\..\refs\omnix\src\Omnix.Serialization.RocketPack\Omnix.Serialization.RocketPack.csproj" />
+    <ProjectReference Include="..\..\refs\omnix\src\Omnix.Serialization\Omnix.Serialization.csproj" />
     <ProjectReference Include="..\Xeus.Messages\Xeus.Messages.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <Service Include="{508349b6-6b84-4df5-91f0-309beebad82d}" />
+    <Compile Remove="excluded_folder\**\*" />
   </ItemGroup>
 </Project>
