@@ -1,13 +1,7 @@
-﻿using Omnix.Base;
-using Omnix.Base.Helpers;
-using Omnix.Cryptography;
-using Omnix.Serialization;
-using Omnix.Serialization.RocketPack;
-using System;
-using System.Buffers;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
+﻿using Omnix.Cryptography;
 using Xeus.Messages;
+
+#nullable enable
 
 namespace Xeus.Messages.Options
 {
@@ -19,33 +13,36 @@ namespace Xeus.Messages.Options
         HttpProxy = 3,
     }
 
-    public sealed partial class ConnectionFilter : RocketPackMessageBase<ConnectionFilter>
+    public sealed partial class ConnectionFilter : Omnix.Serialization.RocketPack.RocketPackMessageBase<ConnectionFilter>
     {
         static ConnectionFilter()
         {
             ConnectionFilter.Formatter = new CustomFormatter();
+            ConnectionFilter.Empty = new ConnectionFilter(string.Empty, (ConnectionType)0, string.Empty);
         }
+
+        private readonly int __hashCode;
 
         public static readonly int MaxSchemeLength = 256;
         public static readonly int MaxProxyUriLength = 1024;
 
         public ConnectionFilter(string scheme, ConnectionType connectionType, string proxyUri)
         {
-            if (scheme is null) throw new ArgumentNullException("scheme");
-            if (scheme.Length > 256) throw new ArgumentOutOfRangeException("scheme");
-            if (proxyUri is null) throw new ArgumentNullException("proxyUri");
-            if (proxyUri.Length > 1024) throw new ArgumentOutOfRangeException("proxyUri");
+            if (scheme is null) throw new System.ArgumentNullException("scheme");
+            if (scheme.Length > 256) throw new System.ArgumentOutOfRangeException("scheme");
+            if (proxyUri is null) throw new System.ArgumentNullException("proxyUri");
+            if (proxyUri.Length > 1024) throw new System.ArgumentOutOfRangeException("proxyUri");
 
             this.Scheme = scheme;
             this.ConnectionType = connectionType;
             this.ProxyUri = proxyUri;
 
             {
-                var hashCode = new HashCode();
-                if (this.Scheme != default) hashCode.Add(this.Scheme.GetHashCode());
-                if (this.ConnectionType != default) hashCode.Add(this.ConnectionType.GetHashCode());
-                if (this.ProxyUri != default) hashCode.Add(this.ProxyUri.GetHashCode());
-                _hashCode = hashCode.ToHashCode();
+                var __h = new System.HashCode();
+                if (this.Scheme != default) __h.Add(this.Scheme.GetHashCode());
+                if (this.ConnectionType != default) __h.Add(this.ConnectionType.GetHashCode());
+                if (this.ProxyUri != default) __h.Add(this.ProxyUri.GetHashCode());
+                __hashCode = __h.ToHashCode();
             }
         }
 
@@ -53,10 +50,10 @@ namespace Xeus.Messages.Options
         public ConnectionType ConnectionType { get; }
         public string ProxyUri { get; }
 
-        public override bool Equals(ConnectionFilter target)
+        public override bool Equals(ConnectionFilter? target)
         {
-            if ((object)target == null) return false;
-            if (Object.ReferenceEquals(this, target)) return true;
+            if (target is null) return false;
+            if (object.ReferenceEquals(this, target)) return true;
             if (this.Scheme != target.Scheme) return false;
             if (this.ConnectionType != target.ConnectionType) return false;
             if (this.ProxyUri != target.ProxyUri) return false;
@@ -64,58 +61,62 @@ namespace Xeus.Messages.Options
             return true;
         }
 
-        private readonly int _hashCode;
-        public override int GetHashCode() => _hashCode;
+        public override int GetHashCode() => __hashCode;
 
-        private sealed class CustomFormatter : IRocketPackFormatter<ConnectionFilter>
+        private sealed class CustomFormatter : Omnix.Serialization.RocketPack.IRocketPackFormatter<ConnectionFilter>
         {
-            public void Serialize(RocketPackWriter w, ConnectionFilter value, int rank)
+            public void Serialize(Omnix.Serialization.RocketPack.RocketPackWriter w, ConnectionFilter value, int rank)
             {
-                if (rank > 256) throw new FormatException();
+                if (rank > 256) throw new System.FormatException();
 
-                // Write property count
                 {
-                    int propertyCount = 0;
-                    if (value.Scheme != default) propertyCount++;
-                    if (value.ConnectionType != default) propertyCount++;
-                    if (value.ProxyUri != default) propertyCount++;
-                    w.Write((ulong)propertyCount);
+                    uint propertyCount = 0;
+                    if (value.Scheme != string.Empty)
+                    {
+                        propertyCount++;
+                    }
+                    if (value.ConnectionType != (ConnectionType)0)
+                    {
+                        propertyCount++;
+                    }
+                    if (value.ProxyUri != string.Empty)
+                    {
+                        propertyCount++;
+                    }
+                    w.Write(propertyCount);
                 }
 
-                // Scheme
-                if (value.Scheme != default)
+                if (value.Scheme != string.Empty)
                 {
-                    w.Write((ulong)0);
+                    w.Write((uint)0);
                     w.Write(value.Scheme);
                 }
-                // ConnectionType
-                if (value.ConnectionType != default)
+                if (value.ConnectionType != (ConnectionType)0)
                 {
-                    w.Write((ulong)1);
+                    w.Write((uint)1);
                     w.Write((ulong)value.ConnectionType);
                 }
-                // ProxyUri
-                if (value.ProxyUri != default)
+                if (value.ProxyUri != string.Empty)
                 {
-                    w.Write((ulong)2);
+                    w.Write((uint)2);
                     w.Write(value.ProxyUri);
                 }
             }
 
-            public ConnectionFilter Deserialize(RocketPackReader r, int rank)
+            public ConnectionFilter Deserialize(Omnix.Serialization.RocketPack.RocketPackReader r, int rank)
             {
-                if (rank > 256) throw new FormatException();
+                if (rank > 256) throw new System.FormatException();
 
                 // Read property count
-                int propertyCount = (int)r.GetUInt64();
+                uint propertyCount = r.GetUInt32();
 
-                string p_scheme = default;
-                ConnectionType p_connectionType = default;
-                string p_proxyUri = default;
+                string p_scheme = string.Empty;
+                ConnectionType p_connectionType = (ConnectionType)0;
+                string p_proxyUri = string.Empty;
 
                 for (; propertyCount > 0; propertyCount--)
                 {
-                    int id = (int)r.GetUInt64();
+                    uint id = r.GetUInt32();
                     switch (id)
                     {
                         case 0: // Scheme
@@ -141,59 +142,60 @@ namespace Xeus.Messages.Options
         }
     }
 
-    public sealed partial class ContentsOptions : RocketPackMessageBase<ContentsOptions>
+    public sealed partial class ContentsOptions : Omnix.Serialization.RocketPack.RocketPackMessageBase<ContentsOptions>
     {
         static ContentsOptions()
         {
             ContentsOptions.Formatter = new CustomFormatter();
+            ContentsOptions.Empty = new ContentsOptions();
         }
+
+        private readonly int __hashCode;
 
         public ContentsOptions()
         {
 
             {
-                var hashCode = new HashCode();
-                _hashCode = hashCode.ToHashCode();
+                var __h = new System.HashCode();
+                __hashCode = __h.ToHashCode();
             }
         }
 
 
-        public override bool Equals(ContentsOptions target)
+        public override bool Equals(ContentsOptions? target)
         {
-            if ((object)target == null) return false;
-            if (Object.ReferenceEquals(this, target)) return true;
+            if (target is null) return false;
+            if (object.ReferenceEquals(this, target)) return true;
 
             return true;
         }
 
-        private readonly int _hashCode;
-        public override int GetHashCode() => _hashCode;
+        public override int GetHashCode() => __hashCode;
 
-        private sealed class CustomFormatter : IRocketPackFormatter<ContentsOptions>
+        private sealed class CustomFormatter : Omnix.Serialization.RocketPack.IRocketPackFormatter<ContentsOptions>
         {
-            public void Serialize(RocketPackWriter w, ContentsOptions value, int rank)
+            public void Serialize(Omnix.Serialization.RocketPack.RocketPackWriter w, ContentsOptions value, int rank)
             {
-                if (rank > 256) throw new FormatException();
+                if (rank > 256) throw new System.FormatException();
 
-                // Write property count
                 {
-                    int propertyCount = 0;
-                    w.Write((ulong)propertyCount);
+                    uint propertyCount = 0;
+                    w.Write(propertyCount);
                 }
 
             }
 
-            public ContentsOptions Deserialize(RocketPackReader r, int rank)
+            public ContentsOptions Deserialize(Omnix.Serialization.RocketPack.RocketPackReader r, int rank)
             {
-                if (rank > 256) throw new FormatException();
+                if (rank > 256) throw new System.FormatException();
 
                 // Read property count
-                int propertyCount = (int)r.GetUInt64();
+                uint propertyCount = r.GetUInt32();
 
 
                 for (; propertyCount > 0; propertyCount--)
                 {
-                    int id = (int)r.GetUInt64();
+                    uint id = r.GetUInt32();
                     switch (id)
                     {
                     }
@@ -204,76 +206,79 @@ namespace Xeus.Messages.Options
         }
     }
 
-    public sealed partial class DownloadOptions : RocketPackMessageBase<DownloadOptions>
+    public sealed partial class DownloadOptions : Omnix.Serialization.RocketPack.RocketPackMessageBase<DownloadOptions>
     {
         static DownloadOptions()
         {
             DownloadOptions.Formatter = new CustomFormatter();
+            DownloadOptions.Empty = new DownloadOptions(string.Empty);
         }
+
+        private readonly int __hashCode;
 
         public static readonly int MaxBasePathLength = 1024;
 
         public DownloadOptions(string basePath)
         {
-            if (basePath is null) throw new ArgumentNullException("basePath");
-            if (basePath.Length > 1024) throw new ArgumentOutOfRangeException("basePath");
+            if (basePath is null) throw new System.ArgumentNullException("basePath");
+            if (basePath.Length > 1024) throw new System.ArgumentOutOfRangeException("basePath");
 
             this.BasePath = basePath;
 
             {
-                var hashCode = new HashCode();
-                if (this.BasePath != default) hashCode.Add(this.BasePath.GetHashCode());
-                _hashCode = hashCode.ToHashCode();
+                var __h = new System.HashCode();
+                if (this.BasePath != default) __h.Add(this.BasePath.GetHashCode());
+                __hashCode = __h.ToHashCode();
             }
         }
 
         public string BasePath { get; }
 
-        public override bool Equals(DownloadOptions target)
+        public override bool Equals(DownloadOptions? target)
         {
-            if ((object)target == null) return false;
-            if (Object.ReferenceEquals(this, target)) return true;
+            if (target is null) return false;
+            if (object.ReferenceEquals(this, target)) return true;
             if (this.BasePath != target.BasePath) return false;
 
             return true;
         }
 
-        private readonly int _hashCode;
-        public override int GetHashCode() => _hashCode;
+        public override int GetHashCode() => __hashCode;
 
-        private sealed class CustomFormatter : IRocketPackFormatter<DownloadOptions>
+        private sealed class CustomFormatter : Omnix.Serialization.RocketPack.IRocketPackFormatter<DownloadOptions>
         {
-            public void Serialize(RocketPackWriter w, DownloadOptions value, int rank)
+            public void Serialize(Omnix.Serialization.RocketPack.RocketPackWriter w, DownloadOptions value, int rank)
             {
-                if (rank > 256) throw new FormatException();
+                if (rank > 256) throw new System.FormatException();
 
-                // Write property count
                 {
-                    int propertyCount = 0;
-                    if (value.BasePath != default) propertyCount++;
-                    w.Write((ulong)propertyCount);
+                    uint propertyCount = 0;
+                    if (value.BasePath != string.Empty)
+                    {
+                        propertyCount++;
+                    }
+                    w.Write(propertyCount);
                 }
 
-                // BasePath
-                if (value.BasePath != default)
+                if (value.BasePath != string.Empty)
                 {
-                    w.Write((ulong)0);
+                    w.Write((uint)0);
                     w.Write(value.BasePath);
                 }
             }
 
-            public DownloadOptions Deserialize(RocketPackReader r, int rank)
+            public DownloadOptions Deserialize(Omnix.Serialization.RocketPack.RocketPackReader r, int rank)
             {
-                if (rank > 256) throw new FormatException();
+                if (rank > 256) throw new System.FormatException();
 
                 // Read property count
-                int propertyCount = (int)r.GetUInt64();
+                uint propertyCount = r.GetUInt32();
 
-                string p_basePath = default;
+                string p_basePath = string.Empty;
 
                 for (; propertyCount > 0; propertyCount--)
                 {
-                    int id = (int)r.GetUInt64();
+                    uint id = r.GetUInt32();
                     switch (id)
                     {
                         case 0: // BasePath
